bugfix to behavior at the last line of the term
Previously, trying to do operations that caused calls to
`terminal_move_to_line` when the cursor was at the bottom of the
terminal and the target line was more than 1 line before the bottom of
the terminal would produce undefined behavior.

The reason for this was that the algorithm for `terminal_move_to_line`
used the DO termcap (aka cud or parm_down_cursor), which produces
undefined behavior when the cursor is on the last line. See the manual
here [0], which says:

> `DO', `UP', `LE', `RI'
> Strings of commands to move the cursor n lines down vertically, up
> vertically, or n columns left or right. Do not attempt to move past any
> edge of the screen with these commands; the effect of trying that is
> undefined. Only a few terminal descriptions provide these commands, and
> most programs do not use them.

Now, the `terminal_move_to_line` algorithm falls back to the code it
used in the case of only a single line movement, which is to output a
`\n` character in a loop, correcting the behavior.

[0]: https://www.gnu.org/software/termutils/manual/termcap-1.3/html_chapter/termcap_4.html#SEC23